Get branch from file system (#4)
* Preliminary version

Displays branch. Need to figure out if there's a way to stop it just an the main brancb - probably not.

* Clean up first attempt

* Save state

Not working. Still trying to figure out how to get the branch where I need it.

* Branch being extracted

However, it's done manually now. Needs to be set to run at startup.

* Save working checkpoint

* Ready for pull request.

Should all be working